+++ webapp/react/src/SafeNotes/modals/Note/SendSuccess.js
