+++ webapp/react/src/SafeNotes/shared/ModalMessage.js
