+++ webapp/react/src/SafeNotes/shared/PinnedMessage.js
